Stock-Hits/requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
dnspython==1.16.0
idna==2.9
mysql-connector-python==8.0.19
protobuf==3.6.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
